requirements.txt
streamlit==1.28.1
pdfplumber==0.10.3
openai==1.3.9
google-generativeai==0.3.0
python-dotenv==1.0.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
